refactor(LineController) : SectionController 를 제거하고 LineController 으로 매핑된 메서드 이동
